ml/requirements.txt
torch
torchvision
opencv-python
numpy
matplotlib
scikit-learn
fastapi
uvicorn
python-multipart
kagglehub
pillow
